+++ cmd/telemetrygen/internal/common/config_test.go
@@ -1,5 +1,4 @@
 // Copyright The OpenTelemetry Authors
-// Copyright (c) 2018 The Jaeger Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package metrics
+package common
 
 import (
 	"testing"
