+++ cmd/telemetrygen/internal/traces/config.go
@@ -15,63 +15,17 @@
 package traces
 
 import (
-	"fmt"
-	"strings"
-	"sync"
-	"time"
-
 	"github.com/spf13/pflag"
-	"go.uber.org/atomic"
-	"go.uber.org/zap"
-	"golang.org/x/time/rate"
-)
 
-var (
-	errFormatOTLPAttributes       = fmt.Errorf("value should be of the format key=\"value\"")
-	errDoubleQuotesOTLPAttributes = fmt.Errorf("value should be a string wrapped in double quotes")
+	"github.com/open-telemetry/opentelemetry-collector-contrib/cmd/telemetrygen/internal/common"
 )
 
 // Config describes the test scenario.
 type Config struct {
-	WorkerCount      int
+	common.Config
 	NumTraces        int
 	PropagateContext bool
-	Rate             int64
-	TotalDuration    time.Duration
 	ServiceName      string
-
-	// OTLP config
-	Endpoint           string
-	Insecure           bool
-	UseHTTP            bool
-	Headers            KeyValue
-	ResourceAttributes KeyValue
-}
-
-type KeyValue map[string]string
-
-var _ pflag.Value = (*KeyValue)(nil)
-
-func (v *KeyValue) String() string {
-	return ""
-}
-
-func (v *KeyValue) Set(s string) error {
-	kv := strings.SplitN(s, "=", 2)
-	if len(kv) != 2 {
-		return errFormatOTLPAttributes
-	}
-	val := kv[1]
-	if len(val) < 2 || !strings.HasPrefix(val, "\"") || !strings.HasSuffix(val, "\"") {
-		return errDoubleQuotesOTLPAttributes
-	}
-
-	(*v)[kv[0]] = val[1 : len(val)-1]
-	return nil
-}
-
-func (v *KeyValue) Type() string {
-	return "map[string]string"
 }
 
 // Flags registers config flags.
@@ -96,45 +50,4 @@ func (c *Config) Flags(fs *pflag.FlagSet) {
 	c.ResourceAttributes = make(map[string]string)
 	fs.Var(&c.ResourceAttributes, "otlp-attributes", "Custom resource attributes to use. The value is expected in the format key=\"value\"."+
 		"Flag may be repeated to set multiple attributes (e.g -otlp-attributes key1=\"value1\" -otlp-attributes key2=\"value2\")")
-}
-
-// Run executes the test scenario.
-func Run(c *Config, logger *zap.Logger) error {
-	if c.TotalDuration > 0 {
-		c.NumTraces = 0
-	} else if c.NumTraces <= 0 {
-		return fmt.Errorf("either `traces` or `duration` must be greater than 0")
-	}
-
-	limit := rate.Limit(c.Rate)
-	if c.Rate == 0 {
-		limit = rate.Inf
-		logger.Info("generation of traces isn't being throttled")
-	} else {
-		logger.Info("generation of traces is limited", zap.Float64("per-second", float64(limit)))
-	}
-
-	wg := sync.WaitGroup{}
-	running := atomic.NewBool(true)
-
-	for i := 0; i < c.WorkerCount; i++ {
-		wg.Add(1)
-		w := worker{
-			numTraces:        c.NumTraces,
-			propagateContext: c.PropagateContext,
-			limitPerSecond:   limit,
-			totalDuration:    c.TotalDuration,
-			running:          running,
-			wg:               &wg,
-			logger:           logger.With(zap.Int("worker", i)),
-		}
-
-		go w.simulateTraces()
-	}
-	if c.TotalDuration > 0 {
-		time.Sleep(c.TotalDuration)
-		running.Store(false)
-	}
-	wg.Wait()
-	return nil
 }
